contrib/serving-simulator/tools/profiler/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pyyaml>=6.0.0
requests>=2.31.0
